+++ lua/plugins/config/vim_move.lua
@@ -1,0 +1,3 @@
+return function(use)
+	use({ 'matze/vim-move', event = 'VimEnter' })
+end