+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - **golang:** reenabled goimports
 - **golang:** commented null_ls goimports gofumpt golines
 - **golang:** null_ls added golines
+- **vim_move:** added vim move
 
 
 <a name="v2.4.0"></a>